New Way of Getting Prices
This takes about 30min to run as there is a 5 second delay to not have OSRS stop traffic to the website as seen in previous tests. This give the price and the name of the item by retrieving the JSON files from the OSRS repository. This only prints at the moment and does not store values yet.